[FIX] : article-refactoring / #87
spaceId int형으로 넘겨주도록 변경

+++ app/src/main/java/com/indipage/presentation/articledetail/ArticleDetailFragment.kt
@@ -202,7 +202,7 @@ class ArticleDetailFragment :
 
     private fun openSpaceDetail(spaceId: Long) {
         findNavController().navigate(
-            R.id.action_article_detail_to_space_detail, bundleOf("spaceId" to spaceId)
+            R.id.action_article_detail_to_space_detail, bundleOf("spaceId" to spaceId.toInt())
         )
     }
 }
